OBC: Removal of segment zero
The current OBC segment list includes a "segment zero" which corresponds
to when the segment maps (segnum_u, segnum_v) point to OBC_NONE, which
is set to zero.

The purpose of this "segment zero" seems to be for avoiding invalid
references, so that OBC%segment(OBC%segnum_[uv](i,j)) always returns a
valid result, included when set to OBC_NONE.  This results in reading
and setting values which are unused or unneeded.

This patch replaces these redundant accesses to segment zero with
various conditional blocks for avoiding these accesses.