(*)Fix an indexing bug in int_density_dz_linear
  Corrected a horizontal indexing bug in int_density_dz_linear that caused the
ISOMIP/layer test case to fail.  This bug was first introduced with PR#732 on
March 8, 2018.  This bug fix will change answers with a linear equation of state
and the finite volume pressure gradient force, however it does not change any of
the verified answers in the MOM6-examples regression suite.