+Move find_interfaces to MOM_state_initialization
  Simplified and cleaned up find_interfaces and moved it from MOM_tracer_Z_init
to MOM_state_initialization, reflecting the fact that it is only used from
within MOM_state_initialization and that it has very little to do with tracer
initialization.  The subroutine fast_bisect was inlined into find_interfaces
and the stand-alone version was deleted.  One minor bug that could change
answers was noted but not corrected.  All answers are bitwise identical, but
the previously publicly visible interface to find_interfaces was made private
to the MOM_state_initialization module.