(*)Revised the routine wave_speeds
  Revised wave_speeds to agree with the calculation of the first mode wave speed
in wave_speed, and to also avoid returning speeds that report back uninitialized
values that do not reproduce across processor layouts and are not dimensionally
consistent.  With these revisions, the diagnosed speeds of the modes reproduce
other solutions, reproduce across layouts, are of sensible magnitude and are
demonstrably dimensionally consistent.  This includes a complete rewrite of
tridiag_det, which previously was just wrong.  Also revised the calls to
register_diag_field for the cn1 and cn_mode# diagnostics to include conversion
factors.  All existing MOM6-examples solutions are bitwise identical, although
there would be answer changes in cases where the wave speeds are used in active
parameterizations.