Use similar routines in wave_speed and wave_speeds
  Revised both wave_speed and wave_speeds to use the same tridiag_det routine to
estimate the determinant and its derivative for their Newton's method solver,
making it more obvious where these two functions are using the same techniques.
In turn, tridiag_deg was modified to take arguments that exploit the fact that
only the upper and lower diagonals plus the candidate eigenvalue are needed in
the specific case used here, avoiding the need for extra arrays and array
copies.  The internal routine tdma6 was also modified to expand similar
expressions relating the center, upper and lower diagonals, and switching the
sign convention in the diagonal arguments to avoid extra copies.  All answers
are  bitwise identical.