Fix epipycnal pairing array size declarations
  Corrected the size of 4 arrays used to describe layer pairings in
tracer_epipycnal_ML_diff, to avoid the possibility of segmentation faults
when there are very few interior layers compared with the number of mixed and
buffer layers.  Also corrected a number of spelling errors in comments.  In
runs that previously worked, all answers should be bitwise identical, and they
are identical in all MOM6-examples test cases.