+Revised interface to write_cputime
  Revised the interface to write_cputime, adding one optional argument and
making another intent inout argument optional when it was not going to be reset,
and also adding a flush of an I/O channel, and added a new subroutine,
MOM_write_cputime_end, to be called during cleanup, potentially via
write_cputime.  The write_cputime calls from the solo driver routines have been
modified in accord with these changes, and a final write_cputime call has been
added so that the time used by the whole run is reported.  These changes should
address the problems in MOM6 issue #853, which should be closed once this PR is
merged into dev/gfdl.  All answers are bitwise identical, but there are minor
interface changes and a new publicly visible subroutine.