*Corrected the test whether to use fluxes%psurf
  Changed the test to detect whether the surface pressure has been set for use
in the equation of state calculations in the thermodynamic step from evaluating
whether forces%p_surf is associated to whether fluxes%p_surf is associated.
Because these two pointers often point to the same array, this only changes
answers in the coupled_AM2_LM3_SIS2/Intersperse_ice_1deg test case, but it could
change answer in other cases, depending on how MOM6 is called.  Because this
only appears to change answers for one test case that is not widely used yet,
and does not impact configurations that are used outside of GFDL, the decision
was taken not to introduce a bug flag to preserve the old incorrect answers.