(*)Correct the use of memory macros for arguments
  Corrected a number of memory macros that would cause MOM6 to give the wrong
answers and probably lead to segmentation faults if global indexing were
enabled.  Macros like NIMEM_ are intended to be used to support static or
dynamic memory in the declaration of ALLOCABLE_ arrays, but instead were being
used to define the sizes of arguments in some routines, which would cause these
arrays to start at 1, not G%isd, inside these routines.  All answers in the
MOM6-examples test suite are bitwise identical.