OBC: segment parser refactoring
The segment configuration parser was encountering problems with GFortran
10.2, where the `fields` array update was being removed by the -O2
optimizer.

In fact, the `parse_segment_data_str` function was doing two separate
operations, where the first call would determine the number of fields
and save their names into an array, and the second call would parse the
data contents of each input field.  The presence of optional arguments
were used to effectively select the preferred operation.

It is possible that the presence of these optional arguments was
interfering with optimization, causing the removal of the `fields`
update.

While this is most likely a bug in the GFortran compiler, we address the
problem by instead splitting this function into two independent
functions, which allows us to remove the optional arguments.

When split, the function appears to work correctly under O2
optimization.