Properly set vertex_shear in set_diffusivity_init
  Always call kappa_shear_at_vertex so that the Vertex_shear element of the
control structure for the MOM_set_diffusivity module is always actively set.  It
had previously been initialized only if USE_JACKSON_PARAM is true, which led to
the issues reported in MOM6 issue #1201.  Also clarified or corrected some of
the comments in MOM_kappa_shear.  All answers will be bitwise identical in all
cases that worked previously, and this PR corrects the problems noted in issue
#1201, which can closed once it is accepted.