Moving autoconf tools to /ac directory
Because autoconf can cause fear and anger in some users, the
configuration files have now been delegated to a subdirectory, ac, which
can be invoked from an arbitrary directory.

This required some unconventional constructs in the file, but they have
luckily been relegated to a few introductory statements, and the bulk of
the code is unchanged.  The only pervasive change is the explicit
references to the ac directory, particularly ac/deps.

.testing also had to be updated to accommodate the autoconf changes.