Post tracer concentration when synced
Tracer concentrations were being posted at the same time as the
transport diagnostics. This leads to some unexpected behavior (e.g. SST
and 3D temperature are not consistent) because the concentrations are
potentially being posted at an intermediate state.

In addition to fixing this problem for all tracers, the subroutine has
been renamed to be more descriptive of what kinds of diagnostics might
fall into this category.