Diagnostic for post-transport tracer concentration
Adds a tracer diagnostic for the intermediate state of tracer
concentrations right after horizontal transport (advection and
diffusion) has occurred. For cases where DIABATIC_FIRST = False, this
will definitely not be identical to the tracer concentration diagnostic
that is posted when the MOM6 state is 'synchronized' (e.g. all forcing
terms within a MOM6 timestep have been calculated and applied).