Testing: Warn on new diagnostic in regression
This patch attempts to detect if a new diagnostic has been added, which
was previously interpreted as a regression.

Previously, any diff in chksum_diag between a PR and dev/gfdl was
interpreted as a regression.  This included new diagnostics, which do
not actually change any existing answers.

This patch only raises an error (blocking the merge) if there are
detected changes in dev/gfdl's chksum_diag.

If the diff only reports new lines in chksum_diag from the PR, then it
reports in the log as a WARN and reports as a PASS to the CI (currently
Travis).

This patch also fixes a minor typo in the .restart test ($$* should have
been $*).  Several Make variables using shell-style brackets {} were
also replaced with Make brackets ().