MOM_sum_output: Reproducibility fix to energy
The total kinetic energy per unit mass, used heavily for model
reproducibility tests, was being computed in a non-reproducibile form:

   KE_mass = 0.5 * (u(i-1)**2 + u(i)**2 + v(j-1)**2 + v**2)

which uses an ambiguous order of summation.

This was causing reproducibility errors in rotational tests where u and
v become swapped.

This patch resolves the problem by explicitly summing the u**2 and v**2
interpolations before adding together.

There is an extremely high change that this will change the output of
`ocean.stats`, causing most tests to report a regression.  Actual model
results should not be affected.