+Change default for SPONGE_IDAMP_VAR to enable PR
Changed the default for SPONGE_IDAMP_VAR from "IDAMP" to "Idamp" so that when this variable is used, as it would be with the proposed change, it does not break any existing experiments.  This could change some MOM_parameter_doc files, but will not change any solutions because the variable whose default is being set was not in use.