+Eliminated sfc_bkgnd_mixing
  Merged the calculations in sfc_bkgnd_mixing into calculate_bkgnd_mixing,
changed some of the arguments to calculate_bkgnd_mixing, and moved the
diagnostics in MOM_bkgnd_mixing to MOM_set_diffusivity to reduce the use of
elements of a control structure across modules.  Also made the Kd_lay arguments
to some of the internal subroutines in MOM_set_diffusivity optional. All answers
are bitwise identical.