+Remove 3-d diagnostic arrays in MOM_bkgnd_mixing
  Made a series of changes to MOM_bkgnd_mixing to eliminate the unnecessary use
of 3-d arrays for diagnostics.  These include passing in 2-d slices of memory
for use with the layer and interface diffusivities and interface viscosity to
calculate_bkgnd_mixing, and the elimination of the kd_bkgnd and kv_bkgnd
elements of bkgnd_mixing_cs.  The diagnostics of background diffusivities and
viscosities are still being made available, but they are only stored in a 3-d
array if the diagnostics are in use.  With these changes the contents of the
control structure for MOM_bkgnd_mixing can now be made private.  All answers
are bitwise identical, but there are changes to public interfaces.