+Move diagnostics out of MOM_CVMix_ddiff
  Moved the registration calls for the double diffusivity diagnostics from
MOM_CVMix_ddiff to MOM_set_diffusivity, so there are no longer duplicated
diagnostic registration calls, and there are not 3-d arrays being allocated in
the CVmix double diffusion module.  Also, the contents of the control structure
for MOM_CVMix_ddiff are now private.  There is a new optional argument to
compute_ddiff_coeffs.  All solutions are bitwise identical, but the order of
some entries in the available_diags files may have changed in some cases.