Fixed the diagnostic internal_heat_temp_tendency
  Corrected the calculation of the diagnostics internal_heat_temp_tendency and
internal_heat_h_tendency in the original version of geothermal.  They had been
wrong because they are based on differences between the temperature and
thickness at the end of the subroutine and stored values, but the stored values
were only being saved in a part of the computational domain and set to 0
elsewhere.  This does not impact the model solutions, but it causes the
regression testing of diagnostics to fail in several of the TCs.  All solutions
are bitwise identical.