Proposal for interface/submodule framework layer
This is a potential solution for implementing a more flexible layer
between MOM6 and its underlying framework.

Currently such a framework is strongly hard-coded to FMS in the
`src/framework` directory, but this can create challenges when using
MOM6 in other coupled models which do not have the same operational
requirements as GFDL.

This solution replaces FMS-based implementations of framework
subroutines with interfaces, which define the API of the framework layer
but do not explicit define it.

The implementations are provided in a separate directory, such as
`config_src/fms_layer`, which is configured at build time.  These
implementations are provided as submodules to the parent module
containing the interface.

This should not be considered a final solution, but rather a starting
point for discussion.

*NOTE*: The current mkmf does not support submodule parsing.  Use these
flags in ac/deps/Makefile if you want to test the build:

```
$ export MKMF_URL="https://github.com/marshallward/mkmf.git"
$ export MKMF_COMMIT="submodule_support"
$ make -j
```

Advantages of separation:

- MOM6 code becomes framework-agnostic, and `src/framework` can
  predominantly contain interfaces to framework-specific
  implementations.

- This is not constrained by any particular API.  The MOM framework API
  is fixed, but is implementation is free to call any combination of
  framework functions in any way needed to achieve the same result.

- This is extendible to an arbitrary number of frameworks, and their
  behavior and testing would be completely separate from MOM6
  development.

Advantages of submodules:

- Framework interfaces are explicit.  Any implementation would need to
  follow the interface as specified in the MOM6 source code.

- A `libMOM6` could be constructed without the framework implementation,
  which could be contained in its own

- Stronger logical separation of framework modules and the other modules
  which reference them.  This may potentially reduce the build times and
  the potential for build errors due to outdated module files.