Refactor of diagnoseMLDbyEnergy computation
This patch includes some minor modifications to the MLD calculation in
`diagnoseMLDbyEnergy`, in order to retain consisent numerical results
after optimization.

Discrepancies were observed in the MLD_EN1 diagnostic with -O0 and -O2
with gfortran 10.2.  This patch restores agreement of these results.

Most of the changes are explicitly replacing certain integer
multiplcations as floating point (3 * x => 3. * x), additional
parentheses, and replacement of a cubic operation with explicit
multiplication (x**3 => x*x*x).  Equations were also reformatted for
closer consistency to the style guide.

This may changes answers for the diagnostic for some runs which rely on
the optimized result.