Makefile testing bug fixes
This fixes some errors in .testing/Makefile.  A macro was not escaping
some of its `$` tokens, and the shell commands were not properly wrapped
in a `$(shell .. )` command.

This was incorrectly wiping results from `results` if one test failed
and a subsequent related test passed (e.g. tc2.repro vs tc2.repro.diag)
because the `ls` command was not being invoked and was always reporting
an empty directory.