+Revised interface to differential_diffuse_T_S
  Revised the interface to differential_diffuse_T_S to explicitly pass the
temperatures and salinities being updated and the diffusivities being used to
update them.  Also added an optional argument to set_diffusivity_init to
indicate whether double diffusion is to be applied, and added a new element to
the diabatic_driver control structure to record whether to apply double
diffusion.  The diabatic_driver code was modified to use these new elements. All
answers are bitwise identical.