+Remove visc%Kd_extra_T and visc%Kd_extra_S
  Removed the allocatable arrays Kd_extra_T and Kd_extra_S from the
vertvisc_type, instead using automatically allocated arrays from within the
various diabatic routines.  This changes the order of some entries in the
MOM_parameter_doc files and changes the comments describing DOUBLE_DIFFUSION, in
part due to the presence of inconsistent descriptions in different parts of the
code.  All answers are bitwise identical, but there are changes the contents of
a transparent public type and changes in the MOM_parameter_doc files.