+(*)Reuse ea_s and eb_s for passive tracer mixing
  Reused the variables ea_s and eb_s for passive tracer mixing in ALE mode,
which simplifies the logic around the calls to call_tracer_column_fns().  This
also corrects a bug using an uninitialized array when double diffusion is not
used and MIX_BOUNDARY_TRACERS is explicitly set to false, which could change
answers for some passive tracers, or it could cause runs to fail altogether, but
there are no known cases that exercises this combination of settings.  The new
runtime parameter MIX_BOUNDARY_TRACER_ALE has been added allow the boundary
mixing for passive to be both set and used, correcting another bug.  All answers
are bitwise for all MOM6-examples test cases, but there are new entries in some
MOM_parameter_doc files.