+Better error handling in reset_face_lengths_list
  Added error handling to reset_face_lengths_list to note (and possibly trigger
a fatal error) for any entries in the CHANNEL_LIST_FILE that do not cause any
open face lengths to change. The runtime parameter FATAL_UNUSED_CHANNEL_WIDTHS
determines whether only warnings are issued or whether there should be a fatal
error; the default is false because otherwise a number of the Baltic test cases
that deliberately share files with global cases would fail due to channels
outside of their domains.  This PR addresses MOM6 issue #683, which should be
closed once this PR is merged in, although by default this new code triggers
warnings and not the suggested fatal error.  All answers are bitwise identical,
but there are new entries in some MOM_parameter_doc files.