Autoconf: netCDF flag fixes
Autoconf netCDF flag configuration was broken due to using nc-config for
Fortran flags, and assuming that C and Fortran files (modules,
libraries, etc) were installed in the same directory.

This patch replaces the lower level --includedir and --libdir flags with
the higher level --fflags and --flibs flags, which provide more
processed Fortran output.

One significant issue here is that --flibs combines the directory (-L)
and library (-l) flags into a single output, even though these should be
split when passed to `ld`, with autoconf passing the former to LDFLAGS
and the latter to LIBS.

We resolve this by stripping the -l flags (via sed) and pass the
remaining flags (almost always -L) to LDFLAGS.  We rely on autoconf
macros to assign the -l flags to LIBS.

This should resolve issues where nc-fortran and nf-fortran are installed
in separate directories.  This was detected in a spack-based
environment, where this would be commonplace.