Correct units of integrated salt flux diagnostics
  Corrected the documented units of diagnostics of globally integrated salt flux
diagnostics.  Also removed two declarations of cmor_units that are redundant
with the ordinary units for these diagnostics, and corrected the logical test
for an error message related to lateral boundary diffusion.  All answers are
bitwise identical, but there are changes in the units in available_diags files
and in the metadata of some output files.