+Deprecate the use of G%ke, using GV%ke instead
  Use GV%ke instead of G%ke everywhere in the MOM6 code to get the number of
layers in a configuration.  This required the addition of new vertical_grid_type
arguments to a number of subroutines, and some unused variables were deleted.
This is one of the final steps in a very long-term project to separate the
vertical and horizontal grid types.  All answers are bitwise identical, but
there are new (non-optional) arguments to 51 subroutines.