Autoconf: explicit netCDF directry flags
The --cflags/--fflags arguments of nc-config/nf-config can be volatile,
since they generally represent library install-time flags, so we instead
replace them with --includedir/--libdir flags, which explicitly denote
the installation paths.

The only real issue here is that nf-config does not have a --libdir
flag, so we use $(nf-config --prefix)/lib and hope that this is where
it's installed.