Add option to read EKE via file for MEKE
Eddy kinetic energy can now be read in from a time-varying field
in a file. The interpolated EKE is used in place of MEKE's PDE
based approximation. All the other options that are used to
estimate viscosity, GM coefficient, and tracer eddy diffusivity
are still valid. Note: this feature has not been extensively
tested.