+Added an explicit interface to open_file
  Added an explicit interface for open_file to MOM_io_wrapper.F90.  This version
only includes the optional arguments that are actually used in the MOM6 or SIS2
code, but adds a new optional MOM_domain_type argument.  The optional arguments
to mpp_open that are not being carried over to the new MOM interface pertain to
archaic file formats and will never be used in MOM6.  All answers are bitwise
identical.