+Added diag_axis_init to MOM_diag_manager
  Added the new function diag_axis_init to MOM_diag_manager, with arguments and
argument orders that are more appropriate for use with MOM6 than the version in
the FMS diag_axis module, which it calls in turn.  A MOM_domain_type is one of
the new optional arguments, along with a level of grid refinement and a flag
indicating that the null_axis id should be returned.  Calls to diag_axis_init
throughout the MOM6 code have been made consistent with this new interface.  All
answers are bitwise identical.