+Simplified MOM_domain_infra dependencies
  Eliminated the dependency of MOM_domain_infra.F90 on MOM_array_transform.F90
by explicitly rotatign the maskmap if necessary. Added the new optional argument
coarsen to get_domain_extent and made three other existing arguments optional,
which makes the routine get_domain_extent_dsamp2 redundant, so it was removed.
Also store the domain name in the MOM_domain type, and use this as the default
when cloning one domain type with another.