+Added 7 thin wrapper routines to MOM_io_infra.F90
  Added explicitly documented interfaces to close_file, flux_file,
io_infra_init, io_infra_end, get_file_info, get_file_fields, and get_field_atts
to MOM_io_infra.F90.  Also changed from get_file_atts to get_field_atts for the
routine that is available via MOM_io.F90 and used in MOM_restart, since this
name captures the use of this call and reflects which of the underlying routines
the overloaded interface mpp_get_atts resolves to.  All answers are bitwise
identical, but there is a localized interface name change, and not all of the
optional arguments in the underlying FMS or mpp routines are being made
available for use with MOM6 code.