(*)Corrected bug in field_chksum
  When the field_chksum interfaces was added, it returned integers, but the set
of mpp_chksum routines from FMS it wraps return 64-bit integers, not 32-bit
integers. For small domain sizes (like in the TC tests) this is not a problem,
but for much larger problems (like tides_025), this results in a truncation of
the results and a change in the checksums, which in turn causes the model to
fail if it tries to read a depth-list file.  This commit restores the model
behavior prior to MOM6 PR#1299 (commit# 797b1958..).  All answers are bitwise
identical in cases that worked with the previous version of the code, but some
depth-list files that were created with MOM6 code after that commit may have
incorrect verification checksums that lead to spurious fatal errors and should
be deleted.